Move the encoding to a new messaging process.
The collector now samples data and pushes it to a queue.
The messenger consumes the new messages, encodes the data,
and will eventually attempt to send the data / store it.

Queuing the raw data here is preferable to encoding it and
queuing the serialized data, as the raw data requires much
less space.

Signed-off-by: Brian Bradley <brian.bradley.p@gmail.com>